Refactor code to support fetching from multiple dexs in the future
