+++ packages/composites/the-graph/src/methods/prices/dex/index.ts
@@ -1,0 +1,1 @@
+export * as uniswapSubgraph from "./uniswap"
